refactor sqlite helpers to be general
