+++ heimdallm/bifrosts/sql/sqlite/tests/select/test_functions.py
@@ -1,7 +1,7 @@
 import pytest
 
 from heimdallm.bifrosts.sql import exc
-from heimdallm.bifrosts.sql.sqlite.select.bifrost import SQLBifrost
+from heimdallm.bifrosts.sql.sqlite.select.bifrost import Bifrost
 
 from .utils import PermissiveConstraints
 
@@ -12,7 +12,7 @@ class MyConstraints(PermissiveConstraints):
 
 
 def test_disallowed_in_select():
-    bifrost = SQLBifrost.mocked(MyConstraints())
+    bifrost = Bifrost.mocked(MyConstraints())
 
     query = "select yep(t1.col) from t1"
     bifrost.traverse(query)
@@ -29,7 +29,7 @@ def test_disallowed_in_select():
 
 
 def test_disallowed_in_where():
-    bifrost = SQLBifrost.mocked(MyConstraints())
+    bifrost = Bifrost.mocked(MyConstraints())
 
     query = "select t1.col from t1 where yep(t1.col)=1"
     bifrost.traverse(query)
@@ -41,7 +41,7 @@ def test_disallowed_in_where():
 
 
 def test_disallowed_in_join():
-    bifrost = SQLBifrost.mocked(MyConstraints())
+    bifrost = Bifrost.mocked(MyConstraints())
 
     query = "select t1.col from t1 join t2 on t1.t2_id=t2.id and t2.col=nope()"
     with pytest.raises(exc.IllegalFunction) as e:
@@ -50,7 +50,7 @@ def test_disallowed_in_join():
 
 
 def test_case_insensitive():
-    bifrost = SQLBifrost.mocked(MyConstraints())
+    bifrost = Bifrost.mocked(MyConstraints())
 
     query = "select YEP(t1.col) from t1"
     bifrost.traverse(query)
