+++ heimdallm/bifrosts/sql/sqlite/tests/select/test_select.py
@@ -3,7 +3,7 @@ from typing import Sequence
 import pytest
 
 from heimdallm.bifrosts.sql import exc
-from heimdallm.bifrosts.sql.sqlite.select.bifrost import SQLBifrost
+from heimdallm.bifrosts.sql.sqlite.select.bifrost import Bifrost
 from heimdallm.bifrosts.sql.utils import FqColumn, JoinCondition, RequiredConstraint
 
 from .utils import PermissiveConstraints
@@ -21,14 +21,14 @@ def test_aliased_select_column():
         def select_column_allowed(self, column: FqColumn) -> bool:
             return column.name in {"t2.col"}
 
-    bifrost = SQLBifrost.mocked(MyConstraints())
+    bifrost = Bifrost.mocked(MyConstraints())
     bifrost.traverse(query)
 
 
 def test_unqualified_columns():
     query = "select col from t1"
 
-    bifrost = SQLBifrost.mocked(PermissiveConstraints())
+    bifrost = Bifrost.mocked(PermissiveConstraints())
     with pytest.raises(exc.UnqualifiedColumn) as e:
         bifrost.traverse(query)
     assert e.value.column == "col"
@@ -45,7 +45,7 @@ ORDER BY payment_date DESC
 LIMIT 5;
     """
 
-    bifrost = SQLBifrost.mocked(PermissiveConstraints())
+    bifrost = Bifrost.mocked(PermissiveConstraints())
     with pytest.raises(exc.UnqualifiedColumn) as e:
         bifrost.traverse(query)
     assert e.value.column == "customer_id"
@@ -68,7 +68,7 @@ def test_escapes(query):
         def condition_column_allowed(self, column: FqColumn) -> bool:
             return column.name in {"t1.id", "t2.jid", "t1.jid"}
 
-    bifrost = SQLBifrost.mocked(MyConstraints())
+    bifrost = Bifrost.mocked(MyConstraints())
     bifrost.traverse(query)
 
 
@@ -85,10 +85,10 @@ def test_disallowed_select_column():
         def select_column_allowed(self, fq_column: FqColumn) -> bool:
             return fq_column.name in {"t2.col"}
 
-    bifrost = SQLBifrost.mocked(AllowColumnConstraints())
+    bifrost = Bifrost.mocked(AllowColumnConstraints())
     bifrost.traverse(query)
 
-    bifrost = SQLBifrost.mocked(DenyColumnConstraints())
+    bifrost = Bifrost.mocked(DenyColumnConstraints())
     with pytest.raises(exc.IllegalSelectedColumn) as excinfo:
         bifrost.traverse(query)
 
@@ -111,7 +111,7 @@ def test_required_constraint():
                 )
             ]
 
-    bifrost = SQLBifrost.mocked(RequiredConstraints())
+    bifrost = Bifrost.mocked(RequiredConstraints())
     with pytest.raises(exc.MissingRequiredConstraint) as excinfo:
         bifrost.traverse(query)
 
@@ -144,14 +144,14 @@ def test_required_constraint():
 def test_broken_query(query):
     """tests that we raise an exception when we cannot parse the query"""
 
-    bifrost = SQLBifrost.mocked(PermissiveConstraints())
+    bifrost = Bifrost.mocked(PermissiveConstraints())
     with pytest.raises(exc.InvalidQuery):
         bifrost.traverse(query)
 
 
 def test_select_column_arith():
     query = "select t1.col + 1 as plus_one from t1"
-    bifrost = SQLBifrost.mocked(PermissiveConstraints())
+    bifrost = Bifrost.mocked(PermissiveConstraints())
     bifrost.traverse(query)
 
 
@@ -164,7 +164,7 @@ def test_select_column_arith():
 )
 def test_select_expr(query):
     """select a non-column expression"""
-    bifrost = SQLBifrost.mocked(PermissiveConstraints())
+    bifrost = Bifrost.mocked(PermissiveConstraints())
     bifrost.traverse(query)
 
 
@@ -178,5 +178,5 @@ def test_conflicting_validations():
         def allowed_joins(self) -> Sequence[JoinCondition]:
             return []
 
-    bifrost = SQLBifrost.mocked(MyConstraints())
+    bifrost = Bifrost.mocked(MyConstraints())
     bifrost.traverse(query)
